abstract out timer code from benchmarks
