+++ benchmarks/create/node-tar-file-async.js
@@ -5,14 +5,8 @@ const fs = require('fs')
 process.on('exit', _ => fs.unlinkSync(file))
 
 const tar = require('../..')
-const start = process.hrtime()
+const timer = require('../timer.js')()
 tar.c({
   file: file,
   cwd: cwd
-}, ['']).then(_ => {
-  const end = process.hrtime(start)
-  const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-  const s = Math.round(end[0]*10 + end[1]/1e8)/10
-  const ss = s <= 1 ? '' : ' (' + s + 's)'
-  console.error('%d%s', ms, ss)
-})
+}, ['']).then(timer)
