+++ benchmarks/create/node-tar-stream-async.js
@@ -5,12 +5,6 @@ const fs = require('fs')
 process.on('exit', _ => fs.unlinkSync(file))
 
 const tar = require('../..')
-const start = process.hrtime()
+const timer = require('../timer.js')()
 const c = tar.c({ cwd: cwd }, [''])
-c.pipe(fs.createWriteStream(file)).on('close', _ => {
-  const end = process.hrtime(start)
-  const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-  const s = Math.round(end[0]*10 + end[1]/1e8)/10
-  const ss = s <= 1 ? '' : ' (' + s + 's)'
-  console.error('%d%s', ms, ss)
-})
+c.pipe(fs.createWriteStream(file)).on('close', timer)
