+++ benchmarks/create/pack-async.js
@@ -5,13 +5,7 @@ const fs = require('fs')
 process.on('exit', _ => fs.unlinkSync(file))
 
 const Pack = require('../../lib/pack.js')
-const start = process.hrtime()
+const timer = require('../timer.js')()
 const p = new Pack({ cwd: cwd })
 p.add('').end()
-p.pipe(fs.createWriteStream(file)).on('finish', _ => {
-  const end = process.hrtime(start)
-  const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-  const s = Math.round(end[0]*10 + end[1]/1e8)/10
-  const ss = s <= 1 ? '' : ' (' + s + 's)'
-  console.error('%d%s', ms, ss)
-})
+p.pipe(fs.createWriteStream(file)).on('finish', timer)
