+++ benchmarks/create/pack-sync.js
@@ -5,12 +5,8 @@ const fs = require('fs')
 process.on('exit', _ => fs.unlinkSync(file))
 
 const Pack = require('../../lib/pack.js')
-const start = process.hrtime()
+const timer = require('../timer.js')()
 const p = new Pack.Sync({ cwd: cwd })
 p.add('')
 fs.writeFileSync(file, p.read())
-const end = process.hrtime(start)
-const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-const s = Math.round(end[0]*10 + end[1]/1e8)/10
-const ss = s <= 1 ? '' : ' (' + s + 's)'
-console.error('%d%s', ms, ss)
+timer()
