+++ benchmarks/extract/node-tar-stream-sync.js
@@ -8,13 +8,7 @@ const file = process.argv[2] || path.resolve(__dirname, '../npm.tar')
 
 const fs = require('fs')
 const tar = require('../..')
-const start = process.hrtime()
+const timer = require('../timer.js')()
 const x = tar.x({ sync: true, cwd: cwd })
-x.on('end', _ => {
-  const end = process.hrtime(start)
-  const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-  const s = Math.round(end[0]*10 + end[1]/1e8)/10
-  const ss = s <= 1 ? '' : ' (' + s + 's)'
-  console.error('%d%s', ms, ss)
-})
+x.on('end', timer)
 x.end(fs.readFileSync(file))
