+++ benchmarks/extract/old-sync.js
@@ -9,14 +9,8 @@ const file = process.argv[2] || path.resolve(__dirname, '../npm.tar')
 const fs = require('fs')
 const Extract = require('tar').Extract
 const data = fs.readFileSync(file)
-const start = process.hrtime()
+const timer = require('../timer.js')()
 const x = new Extract({ path: cwd })
 x.on('entry', entry => entry.resume())
-x.on('close', _ => {
-  const end = process.hrtime(start)
-  const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-  const s = Math.round(end[0]*10 + end[1]/1e8)/10
-  const ss = s <= 1 ? '' : ' (' + s + 's)'
-  console.error('%d%s', ms, ss)
-})
+x.on('close', timer)
 x.end(data)
