+++ benchmarks/extract/tar-fs-async.js
@@ -8,13 +8,7 @@ const file = process.argv[2] || path.resolve(__dirname, '../npm.tar')
 
 const tar = require('tar-fs')
 const fs = require('fs')
-const start = process.hrtime()
+const timer = require('../timer.js')()
 const e = tar.extract(cwd)
-process.on('exit', (code, signal) => {
-  const end = process.hrtime(start)
-  const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-  const s = Math.round(end[0]*10 + end[1]/1e8)/10
-  const ss = s <= 1 ? '' : ' (' + s + 's)'
-  console.error('%d%s', ms, ss)
-})
+process.on('exit', timer)
 fs.createReadStream(file).pipe(e)
