+++ benchmarks/extract/tar-fs-sync.js
@@ -10,13 +10,7 @@ const fs = require('fs')
 const tar = require('tar-fs')
 const data = fs.readFileSync(file)
 
-const start = process.hrtime()
+const timer = require('../timer.js')()
 const e = tar.extract(cwd)
-process.on('exit', (code, signal) => {
-  const end = process.hrtime(start)
-  const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-  const s = Math.round(end[0]*10 + end[1]/1e8)/10
-  const ss = s <= 1 ? '' : ' (' + s + 's)'
-  console.error('%d%s', ms, ss)
-})
+process.on('exit', timer)
 e.end(data)
