+++ benchmarks/extract/unpack-async.js
@@ -9,13 +9,7 @@ const file = process.argv[2] || path.resolve(__dirname, '../npm.tar')
 const Unpack = require('../../lib/unpack.js')
 const fs = require('fs')
 
-const start = process.hrtime()
+const timer = require('../timer.js')()
 const u = new Unpack({ cwd: cwd })
-u.on('close', _ => {
-  const end = process.hrtime(start)
-  const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-  const s = Math.round(end[0]*10 + end[1]/1e8)/10
-  const ss = s <= 1 ? '' : ' (' + s + 's)'
-  console.error('%d%s', ms, ss)
-})
+u.on('close', timer)
 fs.createReadStream(file).pipe(u)
