+++ benchmarks/parse/fast-scan-no-body.js
@@ -4,7 +4,7 @@ const fs = require('fs')
 const data = fs.readFileSync(file)
 const Header = require('../../lib/header.js')
 
-const start = process.hrtime()
+const timer = require('../timer.js')()
 for (let position = 0; position < data.length; position += 512) {
   const h = new Header(data, position)
   if (h.size) {
@@ -12,9 +12,4 @@ for (let position = 0; position < data.length; position += 512) {
     position += blockSize
   }
 }
-
-const end = process.hrtime(start)
-const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
-const s = Math.round(end[0]*10 + end[1]/1e8)/10
-const ss = s <= 1 ? '' : ' (' + s + 's)'
-console.error('%d%s', ms, ss)
+timer()
