+++ benchmarks/timer.js
@@ -1,0 +1,10 @@
+module.exports = _ => {
+  const start = process.hrtime()
+  return _ => {
+    const end = process.hrtime(start)
+    const ms = Math.round(end[0]*1e6 + end[1]/1e3)/1e3
+    const s = Math.round(end[0]*10 + end[1]/1e8)/10
+    const ss = s <= 1 ? '' : ' (' + s + 's)'
+    console.error('%d%s', ms, ss)
+  }
+}
